fix(everything): fix all

+++ src/main/java/polytech/group3/iwa/repositories/LocationRepository.java
@@ -3,4 +3,4 @@ package polytech.group3.iwa.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import polytech.group3.iwa.models.Location;
 
-public interface LocationRepository extends JpaRepository<Location, Long> {}
+public interface LocationRepository extends JpaRepository<Location, Integer> {}
